chore: move injection zones up to appropriate folders
we never use shared anyway.
